// web/src/api/generated/models/SupportGestureKind.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eimir.
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 */
export const SupportGestureKind = {
    KISS: 'KISS',
    CHECK_IN: 'CHECK_IN'
} as const;
export type SupportGestureKind = typeof SupportGestureKind[keyof typeof SupportGestureKind];


export function instanceOfSupportGestureKind(value: any): boolean {
    for (const key in SupportGestureKind) {
        if (Object.prototype.hasOwnProperty.call(SupportGestureKind, key)) {
            if (SupportGestureKind[key as keyof typeof SupportGestureKind] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SupportGestureKindFromJSON(json: any): SupportGestureKind {
    return SupportGestureKindFromJSONTyped(json, false);
}

export function SupportGestureKindFromJSONTyped(json: any, ignoreDiscriminator: boolean): SupportGestureKind {
    return json as SupportGestureKind;
}

export function SupportGestureKindToJSON(value?: SupportGestureKind | null): any {
    return value as any;
}

export function SupportGestureKindToJSONTyped(value: any, ignoreDiscriminator: boolean): SupportGestureKind {
    return value as SupportGestureKind;
}