// web/src/api/generated/models/DashboardView.ts
export interface DashboardView {
    /**
     * 
     * @type {DashboardItem}
     * @memberof DashboardView
     */
    keepsake: DashboardItem | null;
    /**
     * 
     * @type {Array<DashboardItem>}
     * @memberof DashboardView
     */
    recentShared: Array<DashboardItem>;
    /**
     * 
     * @type {DashboardRelationshipDuration}
     * @memberof DashboardView
     */
    relationshipDuration: DashboardRelationshipDuration | null;
    /**
     * 
     * @type {DashboardItem}
     * @memberof DashboardView
     */
    retrospective: DashboardItem | null;
    /**
     * 
     * @type {DashboardSharedStorySummary}
     * @memberof DashboardView
     */
    sharedStorySummary: DashboardSharedStorySummary;
    /**
     * 
     * @type {DashboardSpaceSummary}
     * @memberof DashboardView
     */
    space: DashboardSpaceSummary;
    /**
     * 
     * @type {Date}
     * @memberof DashboardView
     */
    thinkingOfYouAvailableAt: Date | null;
    /**
     * 
     * @type {Array<DashboardItem>}
     * @memberof DashboardView
     */
    upcoming: Array<DashboardItem>;
}